move tableLibs to src/lib
