+++ src/components/Contact/ContactList/components/PageWrapper/components/TableWrapper/TableWrapper.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import EnhancedTable from "./components/EnhancedTable/EnhancedTable";
 import Importer from "./components/Importer";
-import { addRowsFromCsv } from "../../../../tableLibs/dataOperation";
+import { addRowsFromCsv } from "../../../../../../../lib/tableLibs/dataOperation";
 
 class TableWrapper extends React.Component {
   constructor(props) {
