+++ src/components/Contact/ContactList/components/PageWrapper/components/TableWrapper/components/EnhancedTable/components/SelectModal/SelectModal.jsx
@@ -2,13 +2,13 @@ import React, { Component } from "react";
 import "./SelectModal.scss";
 import DropdownList from "./components/DropdownList";
 import StringInput from "./components/StringInput";
-import generateArray from '../../../../../../../../tableLibs/generateArray';
+import generateArray from '../../../../../../../../../../../lib/tableLibs/generateArray';
 import {
   testPhoneNum,
   testEmailAddr,
   testEmptyString,
   testDate,
-} from "../../../../../../../../tableLibs/validation";
+} from "../../../../../../../../../../../lib/tableLibs/validation";
 
 /* ======================Define all Options for Dropdown======================================= */
 const MODAL = {
