+++ src/lib/tableLibs/dataOperation.js
@@ -1,5 +1,5 @@
 import React from "react";
-import JumpButton from "../components/PageWrapper/components/TableWrapper/components/EnhancedTable/components/JumpButton";
+import JumpButton from "../../components/Contact/ContactList/components/PageWrapper/components/TableWrapper/components/EnhancedTable/components/JumpButton";
 
 
 const addRowsFromCsv = (newData) => {
